amend/update Working properly
update_fields removed - this just won't work as we're saving new rows.
If it's truly needed we can use it as a list of fields to copy from self
to old_self to limit what gets saved, but the query efficiency it was
created for will be lost.

amend() now writes three rows again. This is the correct behavior

Test setup fixed to have consistant transaction time.

[#63421322]